Updated rocm-sources ebuilds that ensures kernel files with executable attributes remain (e.g., so "make menuconfig" etc. will work). (These ebuilds make use of the kernel-2 eclass.) Each ebuild exists in its own slot and when the ebuild is removed no files are deleted from /usr/src (i.e., which is the same behaviour of gentoo-sources).
To use with other kernels, simply copy the ebuild renaming the file's
version and set CKV in the ebuild file to be the actual kernel release
of the repository. The latter can be found at the top of the repository's
root Makefile. Use Gentoo conventions for formatting such, e.g., 5.9.0_rc2.